docs: #EN-22, documentation internal linker
