+++ package.json
@@ -75,9 +75,7 @@
     "turbo": "2.2.3",
     "typescript": "5.3.3",
     "vite": "4.5.1",
-    "vite-plugin-dts": "3.7.0",
-    "msw": "2.3.1",
-    "msw-storybook-addon": "2.0.3"
+    "vite-plugin-dts": "3.7.0"
   },
   "packageManager": "pnpm@9.4.0",
   "engines": {
