+++ packages/tiptap/extensions/extension-mathjax/package.json
@@ -2,9 +2,6 @@
   "name": "@edifice-tiptap-extensions/extension-mathjax",
   "version": "1.5.22-develop-pedago.12",
   "description": "Rich Text Editor MathJax Extension",
-  "main": "dist/index.cjs",
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
   "type": "module",
   "exports": {
     ".": {
@@ -12,6 +9,9 @@
       "require": "./dist/index.cjs"
     }
   },
+  "main": "dist/index.cjs",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
   "files": [
     "dist",
     "version.txt"
@@ -20,12 +20,12 @@
     "build": "pnpm run clean && rollup -c",
     "clean": "rm -rf dist",
     "dev": "pnpm run clean && rollup -c -w",
+    "fix": "eslint src --fix --ext ts --report-unused-disable-directives --max-warnings 0",
     "format": "pnpm run format:write && pnpm run format:check",
     "format:check": "prettier --check src/**/*.ts",
     "format:write": "prettier --write src/**/*.ts",
     "generate-commit-version": "node ../../scripts/version.cjs",
     "lint": "eslint src --ext ts --report-unused-disable-directives --max-warnings 0",
-    "fix": "eslint src --fix --ext ts --report-unused-disable-directives --max-warnings 0",
     "release": "pnpm run generate-commit-version && pnpm run build"
   },
   "devDependencies": {
